separate file for TextureAttachment
